Control.Monad.HTTPResponse

Description

Experimenting with using my own pure non-yesod monad for the handlers.

Yesod is just too complicated when it comes to figuring out how to do the really simple stuff that it just seems easier to get out of the yesod code and implement the simple stuff myself instead of digging through the yesod code.

This will probably help with unit testing too

NOTE: An alternative construction of a Pure Monad to help with Proofs and Testing could be to use a Free Monad, I'll try this next time...

Documentation

data Method

Constructors

Get

Post

Put

Delete

Auxiliary Types

newtype CookieMap

Constructors

CookieMap

Fields

cookieMap :: Map Text Text

newtype HTTPHeaders

Constructors

HTTPHeaders

Fields

runHeaders :: Map HeaderName ByteString

data HTTPRequest

Constructors

HTTPRequest

Fields

getPrms :: Map Text Text

getCookieVals :: CookieMap

reqHeaders :: HTTPHeaders

getContent :: ByteString

data HTTPError

The different error codes I have used so far

Constructors

HTTPNotFound

HTTPServerError Text

HTTPBadMethod

HTTPPermissionDenied Text

HTTPInvalidArgs [Text]

data HTTPLogMessage

Constructors

Debug Text

Info Text

data HTTPResponseState

Represents the accrued headers and cookie key-value settings from a REST computation

Constructors

HTTPResponseState

Fields

getAllHeaders :: HTTPHeaders

getSetCookie :: CookieMap

getLogMsgs :: [HTTPLogMessage]

The Monad

type HTTPResponse a = HTTPResponseT Identity a

data HTTPResponseT m a

we want error reporting, logging, accumulation of headers and cookie values, and its going to read in the original request that spurred off the original REST computation.

Writer Helpers

addHeaders :: Monad m => [(HeaderName, ByteString)] -> HTTPResponseT m ()

Sets a header

>>> HTTPReport $ snd $ testHTTPResponse (addHeaders ["Content-Type" .: "application/xml", "Some-Other-Header" .: "And its value"])
Headers:
"Content-Type": "application/xml"
"Some-Other-Header": "And its value"
Session Values:

logMsg :: Monad m => HTTPLogMessage -> HTTPResponseT m ()

For writing log messages...

>>> let debugLogLine = logMsg (Debug "Test Debug Log Line")
>>> let infoLogLine  = logMsg (Info "Test Info Log Line")

>>> (HTTPLogReport . snd . testHTTPResponse) debugLogLine
(Debug) "Test Debug Log Line"

>>> (HTTPLogReport . snd . testHTTPResponse) infoLogLine
(Info) "Test Info Log Line"

>>> (HTTPLogReport . snd . testHTTPResponse . sequence_) [debugLogLine, infoLogLine]
(Debug) "Test Debug Log Line"
(Info) "Test Info Log Line"

>>> (HTTPLogReport . snd . testHTTPResponse . sequence_) [infoLogLine, debugLogLine]
(Info) "Test Info Log Line"
(Debug) "Test Debug Log Line"

putSessionValues :: Monad m => [(Text, Text)] -> HTTPResponseT m ()

Writes session values

>>> HTTPReport $ snd $ testHTTPResponse (putSessionValues ["_id" .: "blah234356d", "somethingelse" .: "blach"])
Headers:
Session Values:
"_id": "blah234356d"
"somethingelse": "blach"

Unwrapping the monad

runHTTPResponse :: HTTPRequest -> HTTPResponse a -> (Either HTTPError a, HTTPResponseState)

runHTTPResponseT :: Monad m => HTTPRequest -> HTTPResponseT m a -> m (Either HTTPError a, HTTPResponseState)

Unit Testing the Pure HTTPResponse Monad

newtype HTTPReport

Wrap a HTTPResponseState with this in a doctest and it will result in the header and cookie values being displayed.

Constructors

HTTPReport HTTPResponseState

newtype HTTPLogReport

Wrap a HTTPResponseState with this in a doctest and it will result in the log lines associated with the state being displayed.

Constructors

HTTPLogReport HTTPResponseState