Foundation

Documentation

data App

The site argument for your application. This can be a good place to keep settings and values requiring initialization before your application starts running, such as database connections. Every handler will have access to the data present here.

Constructors

App

Fields

settings :: AppConfig DefaultEnv Extra

getStatic :: Static

Settings for static file serving.

connPool :: PersistConfigPool PersistConf

Database connection pool.

httpManager :: Manager

persistConfig :: PersistConf

appLogger :: Logger

data AppMessage

Constructors

MsgHello

type Handler = HandlerT App IO

type Widget = WidgetT App IO ()

resourcesApp :: [ResourceTree String]

type Form x = Html -> MForm (HandlerT App IO) (FormResult x, Widget)

getExtra :: Handler Extra

Get the Extra value, used to hold data from the settings.yml file.