Web.ZYesod

Description

Experimenting with using my own non-yesod monad for the handlers.

Yesod is just too complicated when it comes to figuring out how to do the really simple stuff that it just seems easier to get out of the yesod code and implement the simple stuff myself instead of digging through the yesod code.

This module will be about translating to and from the HTTPResponse Monad

Documentation

class (Yesod m, YesodPersist m, PersistUnique (YesodPersistBackend m (HandlerT m IO)), PersistQuery (YesodPersistBackend m (HandlerT m IO)), PersistMonadBackend (YesodPersistBackend m (HandlerT m IO)) ~ SqlBackend) => ZYesod m

convertHTTPResponseIO :: ZYesod m => HTTPResponseT IO a -> HandlerT m IO a

logZMsg :: HTTPLogMessage -> HandlerT m IO ()